+++ cmd/neofs-ir/defaults.go
@@ -50,7 +50,6 @@ func defaultConfiguration(cfg *viper.Viper) {
 	cfg.SetDefault("morph.dial_timeout", 15*time.Second)
 	cfg.SetDefault("morph.validators", []string{})
 
-	cfg.SetDefault("mainnet.endpoint.client", []string{})
 	cfg.SetDefault("mainnet.dial_timeout", 15*time.Second)
 
 	cfg.SetDefault("wallet.path", "")     // inner ring node NEP-6 wallet