+++ cmd/neofs-node/config/morph/config_test.go
@@ -7,7 +7,6 @@ import (
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-node/config"
 	morphconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/morph"
 	configtest "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/test"
-	"github.com/nspcc-dev/neofs-node/pkg/morph/client"
 	"github.com/stretchr/testify/require"
 )
 
@@ -15,26 +14,19 @@ func TestMorphSection(t *testing.T) {
 	t.Run("defaults", func(t *testing.T) {
 		empty := configtest.EmptyConfig()
 
-		require.Panics(t, func() { morphconfig.RPCEndpoint(empty) })
+		require.Panics(t, func() { morphconfig.Endpoints(empty) })
 		require.Equal(t, morphconfig.DialTimeoutDefault, morphconfig.DialTimeout(empty))
 		require.Equal(t, morphconfig.CacheTTLDefault, morphconfig.CacheTTL(empty))
-		require.Equal(t, morphconfig.SwitchIntervalDefault, morphconfig.SwitchInterval(empty))
 	})
 
 	const path = "../../../../config/example/node"
 
-	var (
-		rpcs = []client.Endpoint{
-			{"wss://rpc1.morph.fs.neo.org:40341/ws", 1},
-			{"wss://rpc2.morph.fs.neo.org:40341/ws", 2},
-		}
-	)
+	rpcs := []string{"wss://rpc1.morph.fs.neo.org:40341/ws", "wss://rpc2.morph.fs.neo.org:40341/ws"}
 
 	var fileConfigTest = func(c *config.Config) {
-		require.Equal(t, rpcs, morphconfig.RPCEndpoint(c))
+		require.Equal(t, rpcs, morphconfig.Endpoints(c))
 		require.Equal(t, 30*time.Second, morphconfig.DialTimeout(c))
 		require.Equal(t, 15*time.Second, morphconfig.CacheTTL(c))
-		require.Equal(t, 3*time.Minute, morphconfig.SwitchInterval(c))
 	}
 
 	configtest.ForEachFileType(path, fileConfigTest)